# msps/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "msps",
        "title": "msps",
        "contact": {},
        "version": "latest"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/c/email/send": {
            "post": {
                "description": "处理来自Client的邮件发送请求",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "邮件发送处理",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮件请求参数，作为formData的'data'字段传递",
                        "name": "data",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "邮件附件，作为formData的'attachments'字段传递（可选）",
                        "name": "attachments",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"msg\":\"\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "{\"success\":false,\"msg\":\"请求参数错误\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "{\"success\":false,\"msg\":\"用户未登录\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "403": {
                        "description": "{\"success\":false,\"msg\":\"访问受限\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "404": {
                        "description": "{\"success\":false,\"msg\":\"路径不存在\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "{\"success\":false,\"msg\":\"Internal Server Error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/c/email/{id}/verify": {
            "post": {
                "description": "处理来自Client的邮件确认请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "邮件发送结果确认",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮件唯一标识",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"msg\":\"\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "{\"success\":false,\"msg\":\"请求参数错误\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "{\"success\":false,\"msg\":\"用户未登录\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "403": {
                        "description": "{\"success\":false,\"msg\":\"访问受限\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "404": {
                        "description": "{\"success\":false,\"msg\":\"路径不存在\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "{\"success\":false,\"msg\":\"Internal Server Error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.Response": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "请求结果描述",
                    "type": "string"
                },
                "payload": {
                    "description": "请求结果数据"
                },
                "success": {
                    "description": "是否请求成功",
                    "type": "boolean"
                }
            }
        },
...
    }
}